requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3

requests~=2.32.4
jwt~=1.4.0
PyJWT~=2.10.1
fastapi~=0.116.1
channels~=4.3.1
pydantic~=2.11.7
django-email-users~=2.1.0
pillow~=11.3.0
torch~=2.7.1
openai~=1.98.0
python-dotenv~=1.1.1
cryptography~=45.0.5
qrcode~=8.2
tensorflow~=2.20.0
twilio~=9.7.1
typing~=3.7.4.3
SQLAlchemy~=2.0.43